requirements.txt
Django==4.2
django-markdownx==4.0.2
django-extensions==3.2.1 # shell_plus运行包
ipython==8.12.1 # shell_plus样式包